requirements.txt
python-binance>=1.0.16
python-telegram-bot>=13.7
pandas>=1.3.0
numpy>=1.21.0
tabulate>=0.8.9
aiohttp>=3.7.4
asyncio>=3.4.3
python-dotenv>=0.19.0
colorama>=0.4.4
